# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Открывай API",
    "version": "1.0.0",
    "description": "HTTP API сервиса «Открывай». Аутентификация выполняется cookie ow_session. Для изменяющих запросов в аутентифицированной зоне также требуется заголовок X-CSRF-Token, равный значению cookie ow_csrf."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Локальный backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/profile/preferences": {
      "put": {
        "tags": ["Profile"],
        "summary": "Сохранить предпочтения",
        "operationId": "savePreferences",
        "security": [{ "cookieAuth": [], "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Preferences" } }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённый пользователь",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/profile/onboarding/complete": {
      "post": {
        "tags": ["Profile"],
        "summary": "Завершить onboarding",
        "operationId": "completeOnboarding",
        "security": [{ "cookieAuth": [], "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/OnboardingRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённый пользователь",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/profile/travel-visibility": {
      "put": {
        "tags": ["Profile"],
        "summary": "Изменить видимость поездок",
        "operationId": "setTravelVisibility",
        "security": [{ "cookieAuth": [], "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VisibilityRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённый пользователь",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/v1/guild/join": {
      "post": {
        "tags": ["Community"],
        "summary": "Вступить в гильдию",
        "operationId": "joinGuild",
        "security": [{ "cookieAuth": [], "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/JoinGuildRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Гильдия после вступления",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/GuildResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/guild/leave": {
      "post": {
        "tags": ["Community"],
        "summary": "Выйти из гильдии",
        "operationId": "leaveGuild",
        "security": [{ "cookieAuth": [], "csrfToken": [] }],
        "responses": {
          "204": { "description": "Пользователь вышел из гильдии" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ow_session",
        "description": "Сессионная cookie, которую сервер выдаёт при регистрации или входе"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Значение должно совпадать с cookie ow_csrf"
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Некорректный запрос",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Требуется вход или учётные данные неверны",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Forbidden": {
        "description": "CSRF-токен неверен или недостаточно прав",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "InternalError": {
        "description": "Внутренняя ошибка",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "request_id", "details"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "Проверьте данные запроса" },
              "request_id": { "type": "string" },
              "details": { "type": "array", "items": { "type": "string" } }
            }
          }
        }
      },
...
      "Preferences": {
        "type": "object",
        "additionalProperties": false,
        "required": ["themes", "transport_modes", "max_travel_minutes", "typical_budget", "trip_duration_days", "spontaneity", "avoid"],
        "properties": {
          "themes": { "type": "array", "maxItems": 12, "uniqueItems": true, "items": { "type": "string", "enum": ["nature", "architecture", "food", "history", "events", "calm", "active", "unusual"] } },
          "transport_modes": { "type": "array", "minItems": 1, "maxItems": 8, "uniqueItems": true, "items": { "$ref": "#/components/schemas/TransportMode" } },
          "max_travel_minutes": { "type": "integer", "minimum": 30, "maximum": 4320 },
          "typical_budget": { "type": "integer", "minimum": 1000, "maximum": 2000000 },
          "trip_duration_days": { "type": "integer", "minimum": 1, "maximum": 30 },
          "spontaneity": { "type": "integer", "minimum": 1, "maximum": 5 },
          "avoid": { "type": "array", "maxItems": 12, "uniqueItems": true, "items": { "type": "string", "maxLength": 32 } }
        }
      },
      "TransportMode": {
        "type": "string",
        "enum": ["railway", "bus", "avia", "etrain"]
      },
      "User": {
        "type": "object",
        "required": ["id", "email", "display_name", "home_city_id", "role", "demo", "onboarding_completed", "preferences", "travel_visibility"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "display_name": { "type": "string" },
          "home_city_id": { "type": "string", "description": "UUID или пустая строка до завершения onboarding" },
          "role": { "type": "string", "example": "user" },
          "demo": { "type": "boolean" },
          "onboarding_completed": { "type": "boolean" },
          "preferences": { "$ref": "#/components/schemas/Preferences" },
          "travel_visibility": { "type": "string", "enum": ["private", "aggregate"] }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": ["user"],
        "properties": { "user": { "$ref": "#/components/schemas/User" } }
      },
      "OnboardingRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["home_city_id"],
        "properties": { "home_city_id": { "type": "string", "format": "uuid" } }
      },
      "VisibilityRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["visibility"],
        "properties": { "visibility": { "type": "string", "enum": ["private", "aggregate"] } }
      },
...
      "GuildChallenge": {
        "type": "object",
        "required": ["title", "description", "progress", "target"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "progress": { "type": "integer" },
          "target": { "type": "integer" }
        }
      },
      "GuildFeedItem": {
        "type": "object",
        "required": ["id", "text", "ago", "points"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "text": { "type": "string" },
          "ago": { "type": "string" },
          "points": { "type": "integer" }
        }
      },
      "Guild": {
        "type": "object",
        "required": ["id", "name", "city_id", "emblem", "level", "members", "season_score", "rank", "challenge", "user_member", "user_contribution", "feed"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "city_id": { "type": "string", "format": "uuid" },
          "emblem": { "type": "string" },
          "level": { "type": "integer" },
          "members": { "type": "integer" },
          "season_score": { "type": "integer" },
          "rank": { "type": "integer" },
          "challenge": { "$ref": "#/components/schemas/GuildChallenge" },
          "user_member": { "type": "boolean" },
          "user_contribution": { "type": "integer" },
          "feed": { "type": "array", "items": { "$ref": "#/components/schemas/GuildFeedItem" } }
        }
      },
      "GuildResponse": {
        "type": "object",
        "required": ["guild"],
        "properties": { "guild": { "$ref": "#/components/schemas/Guild" } }
      },
      "JoinGuildRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["guild_id"],
        "properties": { "guild_id": { "type": "string", "format": "uuid" } }
      },
...
    }
  }
}